Create endpoint API
