+++ app/models/message.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 class Message < ApplicationRecord
+  validates :greeting, presence: true
 end
