+++ config/routes.rb
@@ -1,9 +1,12 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  resources :messages
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
+  namespace :api do
+    namespace :v1 do
+      get '/messages/greetings/random', to: 'messages#random'
+    end
+  end
   # Defines the root path route ("/")
   # root "articles#index"
 end
